+++ src/blocks/aside/edit.js
@@ -8,6 +8,7 @@ import classnames from 'classnames';
 // Internal dependencies.
 import themeOptions from '../../global/theme-options.js';
 import allowedBlocks from '../../components/allowed-blocks';
+import { getColorSlug } from '../../global/color-utils.mjs';
 
 // WordPress dependencies.
 import { __ } from '@wordpress/i18n';
@@ -18,29 +19,8 @@ import {
 	InspectorControls,
 	PanelColorSettings,
 	useBlockProps,
-	getColorObjectByColorValue,
 	getColorObjectByAttributeValues,
 } from '@wordpress/block-editor';
-
-/**
- * When given a color it gets the Color Slug from the themeoptions() color
- * palette defined for the theme.
- *
- * @param {*} color
- * @return {string} The slug of the color.
- */
-const getColorSlug = ( color ) => {
-	if ( color ) {
-		const colorObject = getColorObjectByColorValue( themeOptions(), color );
-
-		if ( colorObject.slug ) {
-			return colorObject.slug;
-		}
-	} else {
-		console.error( 'Error: no color.slug value found in color object.' ); // eslint-disable-line no-console
-	}
-	return undefined;
-};
 
 export default function Edit( props ) {
 	const { attributes, setAttributes, className, presetTemplate } = props;
@@ -72,7 +52,7 @@ export default function Edit( props ) {
 							onChange: ( value ) =>
 								setAttributes( {
 									themeColor: value
-										? getColorSlug( value )
+										? getColorSlug( value, themeOptions() )
 										: undefined,
 								} ),
 							label: __( 'Theme' ),