+++ src/blocks/button/button.js
@@ -9,6 +9,7 @@ import classnames from 'classnames';
 
 // Internal dependencies.
 import themeOptions from '../../global/theme-options';
+import { getColorSlug } from '../../global/color-utils.mjs';
 import getAllowedFormats from '../../global/allowed-formats';
 import publicationSlug from '../../global/publication-slug';
 import blockIcons from '../../components/block-icons/';
@@ -34,7 +35,6 @@ import {
 	RichText,
 	URLInput,
 	useBlockProps,
-	getColorObjectByColorValue,
 	getColorObjectByAttributeValues,
 } from '@wordpress/block-editor';
 
@@ -55,26 +55,6 @@ const getClasses = ( className, themeColor, icon ) =>
 		[ `icon-navigateright ${ icon }` ]: icon,
 		[ className ]: className,
 	} );
-
-/**
- * When given a color it gets the Color Slug from the themeoptions() color
- * palette defined for the theme.
- *
- * @param {*} color
- * @return {string} The slug of the color.
- */
-const getColorSlug = ( color ) => {
-	if ( color ) {
-		const colorObject = getColorObjectByColorValue( themeOptions(), color );
-
-		if ( colorObject.slug ) {
-			return colorObject.slug;
-		}
-	} else {
-		console.error( 'Error: no color.slug value found in color object.' ); // eslint-disable-line no-console
-	}
-	return undefined;
-};
 
 // Register the block.
 registerBlockType( 'bu/button', {
@@ -164,7 +144,10 @@ registerBlockType( 'bu/button', {
 								onChange: ( value ) =>
 									setAttributes( {
 										themeColor: value
-											? getColorSlug( value )
+											? getColorSlug(
+													value,
+													themeOptions()
+											  )
 											: undefined,
 									} ),
 								label: __( 'Theme' ),