+++ src/blocks/leadin/leadin.js
@@ -13,6 +13,7 @@ import './editor.scss';
 
 // Internal dependencies
 import themeOptions from '../../global/theme-options';
+import { getColorSlug } from '../../global/color-utils.mjs';
 import getAllowedFormats from '../../global/allowed-formats';
 import publicationSlug from '../../global/publication-slug';
 import Background, {
@@ -37,7 +38,6 @@ import {
 	PanelColorSettings,
 	RichText,
 	URLInput,
-	getColorObjectByColorValue,
 	getColorObjectByAttributeValues,
 } from '@wordpress/block-editor';
 import { applyFilters } from '@wordpress/hooks';
@@ -147,26 +147,6 @@ const blockSupports = {
 	className: false,
 	customClassName: false,
 	multiple: false,
-};
-
-/**
- * When given a color it gets the Color Slug from the themeoptions() color
- * palette defined for the theme.
- *
- * @param {*} color
- * @return {string} The slug of the color.
- */
-const getColorSlug = ( color ) => {
-	if ( color ) {
-		const colorObject = getColorObjectByColorValue( themeOptions(), color );
-
-		if ( colorObject.slug ) {
-			return colorObject.slug;
-		}
-	} else {
-		console.error( 'Error: no color.slug value found in color object.' ); // eslint-disable-line no-console
-	}
-	return undefined;
 };
 
 registerBlockType( 'bu/leadin', {
@@ -546,7 +526,10 @@ registerBlockType( 'bu/leadin', {
 								onChange: ( value ) =>
 									setAttributes( {
 										themeColor: value
-											? getColorSlug( value )
+											? getColorSlug(
+													value,
+													themeOptions()
+											  )
 											: undefined,
 									} ),
 								label: __( 'Theme' ),