+++ src/blocks/modal/edit.js
@@ -7,6 +7,7 @@ import classnames from 'classnames';
 
 // Internal dependencies.
 import themeOptions from '../../global/theme-options';
+import { getColorSlug } from '../../global/color-utils.mjs';
 import Background, { BackgroundControls } from '../../components/background';
 import allowedBlocks from '../../components/allowed-blocks';
 import getAllowedFormats from '../../global/allowed-formats';
@@ -23,33 +24,12 @@ import {
 	InnerBlocks,
 	PanelColorSettings,
 	RichText,
-	getColorObjectByColorValue,
 	getColorObjectByAttributeValues,
 	useBlockProps,
 } from '@wordpress/block-editor';
 
 // Only allow images in the background component for this block.
 const allowedMedia = [ 'image' ];
-
-/**
- * When given a color it gets the Color Slug from the themeoptions() color
- * palette defined for the theme.
- *
- * @param {*} color
- * @return {string} The slug of the color.
- */
-const getColorSlug = ( color ) => {
-	if ( color ) {
-		const colorObject = getColorObjectByColorValue( themeOptions(), color );
-
-		if ( colorObject.slug ) {
-			return colorObject.slug;
-		}
-	} else {
-		console.error( 'Error: no color.slug value found in color object.' ); // eslint-disable-line no-console
-	}
-	return undefined;
-};
 
 //const BUEditorialModalEdit = ( props ) => {
 export default function Edit( props ) {
@@ -85,7 +65,7 @@ export default function Edit( props ) {
 							onChange: ( value ) =>
 								setAttributes( {
 									themeColor: value
-										? getColorSlug( value )
+										? getColorSlug( value, themeOptions() )
 										: undefined,
 								} ),
 							label: __( 'Theme' ),