+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Add color-utils module with getColorSlug() utility function. Used in all blocks with themeOptions() support. 
 - Remove each block's themeable SCSS file from being compiled as these aren't needed in the rep, but instead are used by other themes importing this repo via node.
 - Bug fix: Intro Paragraph block's onSplit function wasn't working to add a paragraph block when the user hits enter. 
 - Bug Fix: Intro Paragraph block was retaining the dropcap when switching Block Styles. Had to change how className was used since the old approach from the beta version of Gutenberg conflicts with current attribute and property usage of `className`.